chore: split utils folder between api and scripts
