+++ basedosdados_api/utils.py
@@ -1,8 +1,26 @@
 # -*- coding: utf-8 -*-
+from os import getenv as _getenv
+
 from django.conf import settings
 
 DB_NAME = settings.DATABASES.get("default", {}).get("NAME", "")
 DB_ENGINE = settings.DATABASES.get("default", {}).get("ENGINE", "")
+
+
+def getadmins():
+    """Get admins from environment variable"""
+    admins = getenv("ADMINS")
+    if admins is None:
+        return []
+    return [admin.split(",") for admin in admins.split(";")]
+
+
+def getenv(var, default=None):
+    """Get environment variable or raise exception if not set"""
+    value = _getenv(var, default)
+    if value is None:
+        raise ValueError(f"Environment variable {var} not set")
+    return value
 
 
 def is_remote():
